Tweaked ext4 data partition parameters
Most importantly, don't do lazy initialization. While this slows down
the resize operation on first boot (from a few seconds to 5 minutes
when using a fast 32 GB card), it also eliminates background writes
during the first hours (about 12 hours on 32 GB cards) of use.

Also increased block size to 4K: this reduces the overhead of various
tables, leading to about 500 MB more usable space on a 32 GB card. The
down side is that more space is wasted on small files, but you'd need
a very large amount of small files for that waste to become more than
the savings.

The 4K block size also reduces the number of I/O operations in the
worst case; in the boot loader benchmarks we saw that many small I/O
operations are much slower than a few larger I/O operations even if
the amount of data transferred is equal.